Move save/load manager under utils module
