+++ aider/commands/load.py
@@ -2,7 +2,7 @@ from typing import List
 
 from aider.commands.utils.base_command import BaseCommand
 from aider.commands.utils.helpers import format_command_result
-from aider.save_load_manager import SaveLoadManager
+from aider.commands.utils.save_load_manager import SaveLoadManager
 
 
 class LoadCommand(BaseCommand):
